Create reusable  component for Input
