+++ src/views/Farms/NewUI/FarmPage.tsx
@@ -52,7 +52,7 @@ import SvgIcon from '../../../components/Launchpad/SvgIcon'
 import { getBscScanAddressUrl } from '../../../utils/bscscan'
 import { Cards2, Card2Container, TokenLogo, Badge, LinearBG, PageContainer } from '../components/FarmCards/styles'
 import { RenderSocials } from '../../../components/Launchpad/Logo'
-
+import InputComponent from './InputComponent'
 
 const FlexC = styled(Flex)`
   padding: 2.5rem;
@@ -117,19 +117,6 @@ const TableStyle = styled.div`
       text-overflow: ellipsis;
     }
   }
-`
-
-const StyledDetails = styled(Flex)`
-  width: 100%;
-  flex-direction: column;
-  & > * {
-    justify-content: space-between;
-    flex: 1;
-    & :first-child {
-      color: ${({theme}) => theme.colors.textSubtle};
-    }
-  }
-  
 `
 
 const ChartStyle = styled(Flex)`
@@ -391,91 +378,7 @@ const RenderFarm: React.FC<{ farmID: string; tblColumns: any }> = ({ farmID, tbl
                 <UnlockButton customTitle='Connect Wallet to Stake' />
               </Flex>
              ) : ( 
-             <>
-             <Flex justifyContent='center' style={{width: '100%'}}>
-               <Grid container spacing={{xs: 2, md: 1}} justifyContent='center'>
-                {
-                  ['15', '90', '180', '365'].map((day) => (
-                    <>
-                      <Grid item xs={12} sm={3} md={3}>
-                        <ButtonSM fullWidth onClick={() => setDayDuration(day)} >
-                          {`${day} Days`}
-                        </ButtonSM>
-                      </Grid>
-                    </>
-                  ))
-                }
-               </Grid>
-             </Flex>
-             <StyledDetails>
-               <Flex>
-                 <Text>
-                   APY
-                 </Text>
-                 <Text>
-                   2%
-                 </Text>
-               </Flex>
-               <Flex>
-                 <Text>
-                   Max fine
-                 </Text>
-                 <Text>
-                   60%
-                 </Text>
-               </Flex>
-               <Flex>
-                 <Text>
-                   Max profit (estimated)
-                 </Text>
-                 <Text>
-                   -
-                 </Text>
-               </Flex>
-               <hr style={{width: '100%'}} />
-               <Flex>
-                 <Text>
-                   You staked
-                 </Text>
-                 <Text>
-                   1000 MGG
-                 </Text>
-               </Flex>
-               <Flex>
-                 <Text>
-                   Your balance
-                 </Text>
-                 <Text>
-                   0 MGG
-                 </Text>
-               </Flex>
-               <Flex>
-                 <Text>
-                   Total staked 
-                 </Text>
-                 <Text>
-                   10000.00 MGG
-                 </Text>
-               </Flex>
-             </StyledDetails>
-             <Flex style={{ flex: '0 50%' }}>
-                <Text>Amount</Text>
-              </Flex>
-              <Flex style={{ flex: '0 50%', justifyContent: 'end' }}>
-                <ButtonSM>Deposit Max</ButtonSM>
-              </Flex>
-              <Flex style={{ flex: '0 100%', position: 'relative' }}>
-                <Input style={{ padding: '1.5rem' }} placeholder="0" type="number" min="0" />
-                <div style={{ position: 'absolute', top: '0.7rem', right: '1.5rem' }}>
-                  <Text color={theme.colors.textSubtle}>
-                    {currentFarm.quoteToken.symbol} - {currentFarm.pairToken.symbol}
-                  </Text>
-                </div>
-              </Flex>
-              <Flex style={{ flex: '0 100%', justifyContent: 'center' }}>
-                <Button fullWidth>Stake</Button>
-              </Flex>
-              </> 
+               <InputComponent dayDuration={dayDuration} dayFunction={setDayDuration} stakingType="farms" currentFarm={currentFarm}  />
               )}
               <Flex style={{ flex: '0 100%' }} />
               <Flex style={{ flex: '0 50%' }}>
